Rename Home Table components to match the rest of them
