+++ frontend/components/home/HomeCards.svelte
