+++ frontend/components/home/HomeTable.svelte
@@ -4,15 +4,15 @@
     import { userStore } from '@/stores'
 
     import CharacterTable from '@/components/character-table/CharacterTable.svelte'
-    import GroupHead from './table/GroupHead.svelte'
-    import RowCovenant from './table/RowCovenant.svelte'
+    import GroupHead from './table/HomeTableGroupHead.svelte'
+    import RowCovenant from './table/row/HomeTableRowCovenant.svelte'
     import RowGold from '@/components/character-table/row/Gold.svelte'
     import RowItemLevel from '@/components/character-table/row/ItemLevel.svelte'
     import RowKeystone from '@/components/character-table/row/Keystone.svelte'
-    import RowMountSpeed from './table/RowMountSpeed.svelte'
-    import RowStatuses from './table/RowStatuses.svelte'
-    import RowTorghast from './table/row/Torghast.svelte'
-    import RowUghQuest from './table/RowUghQuest.svelte'
+    import RowMountSpeed from './table/row/HomeTableRowMountSpeed.svelte'
+    import RowStatuses from './table/row/HomeTableRowStatuses.svelte'
+    import RowTorghast from './table/row/HomeTableRowTorghast.svelte'
+    import RowUghQuest from './table/row/HomeTableRowUghQuest.svelte'
     import RowVaultMythicPlus from '@/components/character-table/row/VaultMythicPlus.svelte'
     import RowVaultRaid from '@/components/character-table/row/VaultRaid.svelte'
