+++ frontend/components/home/table/row/HomeTableRowTorghast.svelte
