+++ frontend/components/home/table/row/HomeTableRowUghQuest.svelte
